init.sh updated

+++ init.sh
@@ -4,28 +4,49 @@ CWD=`pwd`
 cd .vim/bundle/Command-T/ruby/command-t
 ruby extconf.rb
 make
+
 cd ~
-ln -sf $CWD/.vimrc .
-ln -sf $CWD/.Xresources .
-ln -sf $CWD/.gvimrc .
-mkdir -p .vim/autoload
-cd .vim/autoload
+
+if [ ! -e .vimrc ]; then
+	ln -sf $CWD/.vimrc
+fi
+
+if [ ! -e .Xresources ]; then
+	ln -sf $CWD/.Xresources
+fi
+
+if [ ! -e .gvimrc ]; then
+	ln -sf $CWD/.gvimrc
+fi
+
+if [ ! -e .dir_colors ]; then
+	ln -s $CWD/misc_submodules/LS_COLORS/LS_COLORS ~/.dir_colors
+fi
+
+if [ ! -e .conkyrc ]; then
+	ln -s $CWD/.conkyrc
+fi
+
+mkdir -p .mplayer
+cd .mplayer
+if [ ! -e config ]; then
+	ln -sf $CWD/.mplayer/config .
+fi
+
+mkdir -p ~/.config/fontconfig
+cd ~/.config/fontconfig
+if [ ! -e fonts.conf ]; then
+	ln -s $CWD/.config/fontconfig/fonts.conf .
+fi
+
+mkdir -p ~/.config/tint2
+cd ~/.config/tint2
+if [ ! -e tintrc ]; then
+	ln -s $CWD/.config/tint2/tintr2rc
+fi
+
+mkdir -p ~/.vim/autoload
+cd ~/.vim/autoload
 ln -sf $CWD/misc_submodules/vim-pathogen/autoload/pathogen.vim .
 cd ~/.vim
 ln -sf $CWD/.vim/bundle .
-cd bundle
-ln -sf $CWD/misc_submodules/solarized/vim-colors-solarized .
-cd ~
-mkdir -p .mplayer
-cd .mplayer
-ln -sf $CWD/.mplayer/config .
-mkdir -p ~/.config/Terminal
-cd ~/.config/Terminal
-cp $CWD/misc_submodules/solarized/xfce4-terminal/dark/terminalrc .
-mkdir -p ~/.local/share/mc/skins
-cd ~/.local/share/mc/skins
-ln -sf $CWD/misc_submodules/mc-solarized-skin/solarized.ini .
-mkdir -p ~/.config/fontconfig
-cd ~/.config/fontconfig
-ln -s $CWD/.config/fontconfig/fonts.conf .
-cp misc_submodules/LS_COLORS/LS_COLORS ~/.dir_colors
